% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foxtrot.R
\name{bachata}
\alias{bachata}
\title{bachata}
\usage{
bachata(.tbl, ..., .name = "data", .env = caller_env())
}
\arguments{
\item{.tbl}{TODO}

\item{...}{TODO}

\item{.name}{TODO}

\item{.env}{TODO}
}
\description{
bachata
}
