% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{bolero}
\alias{bolero}
\title{Filtering rows}
\usage{
bolero(.tbl, ..., .op = and, .env = caller_env())
}
\arguments{
\item{.tbl}{data frame, most likely grouped}

\item{..., .env}{formulas and caller environment}

\item{.op}{binary operator to \code{\link[=reduce]{reduce()}} results when there are multiple \code{...}}
}
\value{
A tibble with matching rows
}
\description{
Filtering rows
}
\examples{
iris \%>\%
  bolero(~ Sepal.Length > 5.5, ~Sepal.Width >= 4)

}
