% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{choreography}
\alias{choreography}
\title{choreography}
\usage{
choreography(.tbl, ..., .formulas = list2(...), .env = caller_env())
}
\arguments{
\item{.tbl}{TODO}

\item{...}{TODO}

\item{.formulas}{TODO}

\item{.env}{TODO}
}
\description{
choreography
}
