% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{madison}
\alias{madison}
\title{madison}
\usage{
madison(.tbl, ..., .name = "data", .env = caller_env())
}
\arguments{
\item{.tbl}{TODO}

\item{...}{TODO}

\item{.name}{TODO}

\item{.env}{TODO}
}
\description{
madison
}
