% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swing.R
\name{twist}
\alias{twist}
\title{twist}
\usage{
twist(.fun, ..., .tbl = get_tbl(), .name = "data",
  .env = caller_env())
}
\arguments{
\item{.fun}{TODO}

\item{...}{TODO}

\item{.tbl}{TODO}

\item{.name}{TODO}

\item{.env}{TODO}
}
\description{
twist
}
