% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{waltz}
\alias{waltz}
\title{waltz() takes a grouped tibble and a list of formulas and returns a tibble with: as many columns as supplied formulas,
one row per group. It does not prepend the grouping variables (see tango for that).}
\usage{
waltz(.tbl, ..., .env = caller_env())
}
\arguments{
\item{.tbl}{a grouped tibble}

\item{...}{a list of formulas}

\item{.env}{?? to do}
}
\value{
a tibble with: as many columns as supplied formulas,
one row per group
}
\description{
waltz() takes a grouped tibble and a list of formulas and returns a tibble with: as many columns as supplied formulas,
one row per group. It does not prepend the grouping variables (see tango for that).
}
\examples{
g <- iris \%>\% group_by(Species)
g \%>\%
waltz(
 Sepal.Length = ~mean(Sepal.Length),
 Sepal.Width  = ~mean(Sepal.Width)
)
}
