% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumba.R
\name{zumba}
\alias{zumba}
\title{zumba}
\usage{
zumba(.var, ..., .tbl = get_tbl(), .name = "data",
  .env = caller_env())
}
\arguments{
\item{.var}{TODO}

\item{...}{TODO}

\item{.tbl}{TODO}

\item{.name}{TODO}

\item{.env}{TODO}
}
\description{
zumba
}
