% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{bolero}
\alias{bolero}
\title{bolero}
\usage{
bolero(.tbl, ..., .op = and, .env = caller_env())
}
\arguments{
\item{.tbl}{TODO}

\item{...}{TODO}

\item{.op}{TODO}

\item{.env}{TODO}
}
\description{
bolero
}
