% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choreography.R
\name{choreography}
\alias{choreography}
\title{choreography}
\usage{
choreography(.tbl, ..., .env = caller_env())
}
\arguments{
\item{.tbl}{A data frame}

\item{...}{A variable number of formulas. \code{choreography()} only
uses the rhs of each of the formulas.}

\item{.env}{parent environment of the created function, see \code{\link[rlang:new_function]{rlang::new_function()}}}
}
\value{
a function that can be called with a single argument that
represents indices.

When called with an integer vector \code{idx}, the function returns a list
of each of the expressions given on the rhs evaluated on the subset
of the columns, i.e. in the formula \code{~mean(Sepal.Length)} the column
\code{Sepal.Length} stands for \code{Sepal.Length[idx]}.
}
\description{
The choreography is a central concept of the dance
package, most of the time you don't need to use it directly, but it is
used by many other functions like \code{\link[=tango]{tango()}}, \code{\link[=samba]{samba()}}, ...
}
\examples{

moves <- choreography(iris,
  Sepal.Length = ~mean(Sepal.Length),
  Sepal.Width  = ~mean(Sepal.Width)
)

moves(1:10)
# this returns the same as
list(
  Sepal.Length = mean(iris$Sepal.Length[1:10]),
  Sepal.Width  = mean(iris$Sepal.Width[1:10])
)

}
