% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{mambo}
\alias{mambo}
\title{mambo}
\usage{
mambo(.fun, ..., .tbl = get_tbl(), .op = and, .env = caller_env())
}
\arguments{
\item{.fun}{TODO}

\item{...}{TODO}

\item{.tbl}{TODO}

\item{.op}{TODO}

\item{.env}{TODO}
}
\description{
mambo
}
