% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tibble}
\alias{\%>\%}
\alias{and}
\alias{or}
\alias{group_by}
\alias{\%<-\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_by}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{and}}, \code{\link[magrittr]{or}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}

  \item{zeallot}{\code{\link[zeallot]{\%<-\%}}}
}}

