% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumba.R
\name{rumba}
\alias{rumba}
\title{rumba}
\usage{
rumba(.var, ..., .tbl = get_tbl(), .name = "{fun}",
  .env = caller_env())
}
\arguments{
\item{.var}{TODO}

\item{...}{TODO}

\item{.tbl}{TODO}

\item{.name}{TODO}

\item{.env}{TODO}
}
\description{
rumba
}
