% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swing.R
\name{swing}
\alias{swing}
\title{swing}
\usage{
swing(.fun, ..., .tbl = get_tbl(), .name = "{var}",
  .env = caller_env())
}
\arguments{
\item{.fun}{TODO}

\item{...}{TODO}

\item{.tbl}{TODO}

\item{.name}{TODO}

\item{.env}{TODO}
}
\description{
swing
}
